feat: Create HTTP server with basic routing and comments
Implemented a simple HTTP server using Node.js.
Added routes for / (Home Page) and /about (About Page).
Included detailed comments in Better Comments style for clarity and future reference.
Highlighted the use of the http module and provided a link to the official documentation.
Introduced a TODO for exploring Express.js as an alternative framework.